+++ CHANGELOG.md
@@ -1,17 +1,49 @@
-### Version 1.02 [XX-05-2020]
+### Version 1.04 [XX-XX-2020]
+**Required ElvUI Version: 11.50**  
+**Changes:**  
+Completely new UI layout for both DPS/TANK and Healing  
+• There will be a Discord post about the changes  
+Removed: GottaGoFast addon profile  
+• Note: I will no longer use GottaGoFast in Shadowlands  
+**New Additions:**  
+New option: Reset Nameplate CVars  
+• This button can reset Nameplate CVars to LuckyoneUI defaults  
+**Misc:**  
+Redesigned options  
+Bigwigs profile update  
+Details profile update  
+ProjectAzilroka profile update  
+Updated Credits  
+___
+### Version 1.03 [24-07-2020]
+**Changes:**  
+Enabled RaidDebuff Indicator for Raid40 frames  
+**Misc:**  
+Fixed Raid40 frames for the healer layout  
+More Credits  
+___
+### Version 1.02 [08-07-2020]
 **Required ElvUI Version: 11.42**  
-
 **Changes:**  
 Set the default texture to Solid  
 Set up the new ElvUI 11.42 DataTexts  
 Updated both chat windows  
 **New Additions:**  
-New installer step: AddOn Profiles  
-(This step will set up BigWigs, Details, ProjectAzilroka)  
+New installer steps: AddOn profiles  
+• BigWigs  
+• Details! Damage Meter  
+• ProjectAzilroka  
+• GottaGoFast  
 New installer step: CVars  
-(This step will set up tweaks like maxCameraDistance)  
+• Set up tweaks like maxCameraDistance  
+New installer step: Style Filters  
+• Set up ElvUI NamePlate Style Filters  
+New fonts:  
+• Alpha Echo  
+• Roboto Light  
 **Misc:**  
-Updated config  
+Updated options  
+Added Discord  
 Added credits  
 ___
 ### Version 1.01 [29-02-2020]